installer/requirements.txt
#######requirements.txt #######

###### sudo pip install --ignore-installed six -r requirements.txt ######

###### Requirements without Version Specifiers ######
requests
psycopg2
numpy
GitPython
lxml
python-magic
libmagic
netifaces
tld
pylint
pyyaml
###### Requirements with Version Specifiers ######
jsonschema==2.5.1
pytest==2.9.2
Cython==0.25.2
pyOpenSSL==16.2.0
backports.functools_lru_cache==1.3
certifi==2017.1.23
ed25519ll==0.6
protobuf==3.2.0
ipaddr==2.1.11
keyring==10.3
ndg_httpsclient==0.4.2
pyasn1==0.2.3
redis==2.10.5
setuptools_svn==1.2
wincertstore==0.2
pyxdg==0.25
coverage==4.3.4
Sphinx==1.4.8
setuptools==27.2.0
